% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{projproj}
\alias{projproj}
\title{Shiny application that shows different projections with "movable" map
rendered by ggplot2 rnaturalearth.}
\usage{
projproj()
}
\description{
Shiny application that shows different projections with "movable" map
rendered by ggplot2 rnaturalearth.
}
